backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5

PyJWT==2.8.0
passlib==1.7.4
email-validator==2.1.1

redis==5.0.1
celery==5.3.4

python-dotenv==1.0.1

marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

requests==2.31.0
